Update README.md Troubleshooting Section
Gave the troubleshooting section a numbered list and added #4 "Ensure that other devices such as capture cards are disconnected so your webcam can be detected."

This advice has personally helped me. I had an issue where the program was not detecting my camera and was showing a black screen. In reality, it was displaying my Elgato capture card. Once I disconnected it and restarted my computer my webcam finally appears.